Move derserialization compiler into internal
